consumer/requirements.txt
azure-core==1.22.1
azure-identity==1.7.1
azure-kusto-data==3.0.1
azure-kusto-ingest==3.0.1
azure-storage-blob==12.9.0
azure-storage-queue==12.1.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
ijson==3.1.4
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
scapy==2.4.5
six==1.16.0
urllib3==1.26.8
fastapi
uvicorn
pydantic
httpx